helper function get_online_cpus() for CPU iteration
Since kernel version 4.9.0 BPF stopped working in a KVM guest.
The problem are calls to perf_event_open with CPU identifiers which do
not exist (ENODEV). The root cause for this is that the current code
assumes ascending numbered CPUs. However, this is not always the case
(e.g. CPU hotplugging).

This patch introduces a get_online_cpus() helper function which returns
a list of online CPUs and uses this function for all iterations over
CPUs.

Signed-off-by: Andreas Gerstmayr <andreas.gerstmayr@catalysts.cc>